+++ crates/polars-lazy/Cargo.toml
@@ -217,7 +217,7 @@ arg_where = ["polars-plan/arg_where"]
 search_sorted = ["polars-plan/search_sorted"]
 merge_sorted = ["polars-plan/merge_sorted"]
 meta = ["polars-plan/meta"]
-pivot = ["polars-core/rows", "polars-ops/pivot"]
+pivot = ["polars-core/rows", "polars-ops/pivot", "polars-plan/pivot"]
 top_k = ["polars-plan/top_k"]
 semi_anti_join = ["polars-plan/semi_anti_join"]
 cse = ["polars-plan/cse", "polars-mem-engine/cse"]
